fix: rename fileSharingServices to .ts
