+++ src/services/filesSharingServices.spec.js
@@ -5,7 +5,7 @@
 import axios from '@nextcloud/axios'
 import { generateOcsUrl } from '@nextcloud/router'
 
-import { shareFile } from './filesSharingServices.js'
+import { shareFile } from './filesSharingServices.ts'
 
 jest.mock('@nextcloud/axios', () => ({
 	post: jest.fn(),
